Style announcements part
The announcements is given a white background so it stands out from
the background. It is also given rounded corners, and some lorem text.
A line is shown in between each post in announcements in order to make
it easier for the user to be able to easily spot each separate post.
The text is given a grey color, and the title is bold.